Backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
async-timeout==4.0.3
attrs==25.4.0
bcrypt==4.1.2
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
dnspython==2.6.1
edge-tts==7.2.7
email-validator==2.1.0
exceptiongroup==1.3.1
fastapi==0.110.0
ffmpeg-python==0.2.0
filelock==3.20.1
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
google-ai-generativelanguage==0.6.4
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.5.4
googleapis-common-protos==1.72.0
groq==1.0.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
langchain-groq
langchain-core
langchain-community
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
motor==3.3.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
numba==0.63.1
numpy<2.0.0
faiss-cpu
openai-whisper==20231117
packaging==25.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.8.0
pymongo==4.6.3
pyparsing==3.3.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.29.0
yarl==1.22.0
